modularizing mcpServer resources, tools, and formatting
